Move screenshots Vue component out into a separate component file for testing.
